requirements.txt
requests>=2.31.0
schedule>=1.2.0
Pillow>=10.0.0
flask>=3.0.0
openai>=1.68.2
python-dotenv>=1.0.1
imageio>=2.37.0
imageio-ffmpeg>=0.6.0
numpy>=2.2.4
